# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  golang-auth-api-mysql_internal_domain.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  golang-auth-api-mysql_internal_domain.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  golang-auth-api-mysql_internal_domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  golang-auth-api-mysql_internal_domain.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  golang-auth-api-mysql_internal_domain.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  golang-auth-api-mysql_internal_domain.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  golang-auth-api-mysql_internal_domain.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  golang-auth-api-mysql_internal_domain.UpdateProductRequest:
    properties:
      description:
        type: string
      name:
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Complete REST API with JWT authentication, role-based access control,
    audit logging, and product management
  title: Golang Auth API
  version: "1.0"
paths:
  /audit-logs:
    get:
      consumes:
      - application/json
      description: Get paginated and filtered audit logs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by resource
        in: query
        name: resource
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data: AuditListResponse'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit logs (Admin only)
      tags:
      - Audit
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset link to user's email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Password reset email sent'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Not Found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get tokens
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: AuthResponse'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: logged out successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: AuthResponse'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Invalid refresh token'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: AuthResponse'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend email verification link to user
      parameters:
      - description: Resend Verification Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Verification email sent'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Not Found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend verification email
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using reset token
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Password reset successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Not Found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Auth
  /auth/verify-email:
    get:
      consumes:
      - application/json
      description: Verify user email address using verification token
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Email verified successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Not Found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email address
      tags:
      - Auth
  /products:
    get:
      consumes:
      - application/json
      description: Get paginated and filtered list of products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: 'data: ProductListResponse'
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Create Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: Product'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create product (Admin only)
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: product deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete product (Admin only)
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get single product details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: Product'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Product not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/golang-auth-api-mysql_internal_domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: Product'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Product not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update product (Admin only)
      tags:
      - Products
...
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data, page, limit, total_items
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: user deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user (Admin only)
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"